// context/Global/Global.types.ts
export type GlobalContextType = {
  isLoggedIn: boolean;
  setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
  user: any;
  setUser: React.Dispatch<React.SetStateAction<any>>;
  isLoading: boolean;
  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
  isDataUpToDate?: boolean;
  setIsDataUpToDate: React.Dispatch<React.SetStateAction<boolean>>;
};